+++ android/src/main/java/com/magicleap/magicscript/scene/nodes/base/UiBaseLayout.kt
@@ -22,6 +22,7 @@ import android.os.Looper
 import com.facebook.react.bridge.ReadableMap
 import com.google.ar.sceneform.Node
 import com.magicleap.magicscript.scene.nodes.layouts.manager.LayoutManager
+import com.magicleap.magicscript.scene.nodes.layouts.params.LayoutParams
 import com.magicleap.magicscript.scene.nodes.props.Bounding
 
 // Base class for layouts (grid, linear, rect)
