+++ android/src/main/java/com/magicleap/magicscript/scene/nodes/layouts/params/PageViewLayoutParams.kt
@@ -1,4 +1,4 @@
-package com.magicleap.magicscript.scene.nodes.base
+package com.magicleap.magicscript.scene.nodes.layouts.params
 
 import com.magicleap.magicscript.scene.nodes.props.Alignment
 import com.magicleap.magicscript.scene.nodes.props.Padding
