+++ android/src/test/java/com/magicleap/magicscript/scene/nodes/layouts/manager/HorizontalLinearLayoutManagerTest.kt
@@ -18,7 +18,7 @@ package com.magicleap.magicscript.scene.nodes.layouts.manager
 
 import com.magicleap.magicscript.NodeBuilder
 import com.magicleap.magicscript.layoutUntilStableBounds
-import com.magicleap.magicscript.scene.nodes.base.LayoutParams
+import com.magicleap.magicscript.scene.nodes.layouts.params.LayoutParams
 import com.magicleap.magicscript.scene.nodes.base.TransformNode
 import com.magicleap.magicscript.scene.nodes.props.Alignment
 import com.magicleap.magicscript.scene.nodes.props.Bounding
@@ -131,11 +131,12 @@ class HorizontalLinearLayoutManagerTest {
         assertEquals(-5f, childrenList[1].localPosition.y, EPSILON)
     }
 
-    private fun getLayoutParams() = LayoutParams(
-        size = size,
-        itemPadding = itemPadding,
-        itemHorizontalAlignment = itemHorizontalAlignment,
-        itemVerticalAlignment = itemVerticalAlignment
-    )
+    private fun getLayoutParams() =
+        LayoutParams(
+            size = size,
+            itemPadding = itemPadding,
+            itemHorizontalAlignment = itemHorizontalAlignment,
+            itemVerticalAlignment = itemVerticalAlignment
+        )
 
 }
